Change RegEx for yield value to match digits in json-ld less restrictive
According to https://schema.org/recipeYield the yield value is specified like "recipeYield": "1 loaf"
The suggested RegEx  /(\d*)/ is a more generic approach to find digits in the recipeYield string instead of  /^.*?(\d*)/

For example this works for chefkoch.de ("recipeYield": "6 Portion(en)") and fittastetic.com ("recipeYield": ["2","2 Portionen"]).
The old RegEx /^.*?(\d*)/ worked on chefkoch.de but not e.g. on fittastetic.com

Because we have the if function in the next line (221) that deals with several occurrences and picks the first one, it should be save to do it that way.

Signed-off-by: ZwoaBier <ichhasseapple@gmail.com>