Remove tight coupling between input and paste
In `handleInput`, check if content was pasted using `event.inputType`.
This uses the event as the source of truth, so that we don't need the
call to `nextTick` nor to read attributes set by `handlePaste`.

In `handlePaste`, if a single line was pasted, emit just like
`handleInput` was doing for a single line pasted.

Remove the attributes `contentPasted` and `singleLinePasted`.

All of this removes the tight coupling between `handleInput` and
`handlePaste`. There is no longer a need for the two attributes
`singleLinePasted` or `contentPasted`.

Signed-off-by: Marcel Robitaille <mail@marcelrobitaille.me>
Signed-off-by: Christian Wolf <github@christianwolf.email>