Overwrite breadcrumbs width properly
Previously, we introduced a class `.breadcrumbs` that sets `width:
calc(100% - 60px)` to compensate for the left margin. However,
`.breadcrumb` defined in `@nextcloud/vue` Breadcrumbs
sets `width: 100%`. This was getting priority over our `calc` rule.
Not wanting to use `!important`, the solution I chose was to move this
single rule to the unscoped style tag and to add additional class name
selectors in order to increase the specificity.

Signed-off-by: Marcel Robitaille <mail@marcelrobitaille.me>