Don't prevent default event of inserting #
When refactoring they keydown handler of EditInputGroup,
a regression was introduced where the # would not be inserted when
linking to another recipe. This is because we previously had a keyup
event listener, which cannot prevent the key from being insterted, but
now we are using a keydown listener, which can. Therefore, the
`e.preventDefault()` that was already present in the code now has
unintended side effects. We can remove it, there is no reason to have it
there.

Signed-off-by: Marcel Robitaille <mail@marcelrobitaille.me>