Rename the app from nextcloud-cookbook to cookbook
The old name causes problems when using the webpack `file-loader`. The
URL from this would start `/apps/nextcloud-cookbook` although the proper
install location and the proper URL to load the file is
`/apps/cookbook`.
Besides this `file-loader` issue, this change makes everything more
consistent and more conformal to best practices.

Signed-off-by: Marcel Robitaille <mail@marcelrobitaille.me>